Main GUI Screen Working
-increased min supported ver. to 15
-fixed ButterKnife dependency
-wired up the recycker view for the Main activity
-main activity now reads and parses given file properly now with Gson
-fixed small bugs
-moved given recipe file to raw resource folder
-added an image background to main screen

+++ app/src/main/java/com/learning/twilson/baking/models/Ingredient.java
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName;
 public class Ingredient {
     @SerializedName("quantity")
     @Expose
-    private Integer quantity;
+    private Float quantity;
 
     @SerializedName("measure")
     @Expose
@@ -16,11 +16,11 @@ public class Ingredient {
     @Expose
     private String ingredient;
 
-    public Integer getQuantity() {
+    public Float getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Float quantity) {
         this.quantity = quantity;
     }
